[Refactor] 초대받은 대시보드용 Invited 폴더 생성
